TestCurrency: remove access control
For simplicity and backward compatibility with old Ensuro's
TestCurrency, I removed the access control for the default TestCurrency
contract.

I added TestCurrencyAC for code that was using the version with access
control.

The initCurrency function should switch between one or the other
depending if the extra parameter `admin` was sent.

Also, I removed the dependency between TestERC4626 and TestCurrency,
just in case the project using the library already has a contract called
TestCurrency.